tab's sticky behaviour fixes & refactor
