+++ src/components/Tabs/utils.ts
@@ -1,11 +1,11 @@
-import { activeTabFlagsInterface } from "./types"
+import { ActiveTabFlagsInterface } from "./types"
 
 export const getWidthFromRef = (refNode: React.RefObject<HTMLDivElement>) => {
   return refNode.current ? refNode.current.offsetWidth : 0
 }
 
 export const getTabContainerStyle = (
-  activeTabFlags: activeTabFlagsInterface,
+  activeTabFlags: ActiveTabFlagsInterface,
   activeKeyIndex: number,
   itemIndex: number,
   itemRef: React.RefObject<HTMLDivElement>
